Separate out micro-booter tests (mb_tests) from the init
This is required so that we don't have multiple copies of micro-benchmarking
kernel api in vkernel and micro_booter.
