+++ src/components/implementation/no_interface/vkernel/Makefile
@@ -1,4 +1,4 @@
-C_OBJS=micro_booter.o vkernel.o
+C_OBJS=mb_tests.o micro_booter.o vkernel.o
 ASM_OBJS=cos_asm_scheduler.o inv.o
 COMPONENT=vkernel_boot.o
 INTERFACES=
