+++ src/components/implementation/no_interface/vkernel/mb_tests.c
@@ -1,0 +1,1 @@
+../../tests/micro_booter/mb_tests.c
