+++ src/components/implementation/tests/micro_booter/Makefile
@@ -1,4 +1,4 @@
-C_OBJS=micro_booter.o
+C_OBJS=micro_booter.o mb_tests.o
 ASM_OBJS=cos_asm_scheduler.o inv.o
 COMPONENT=micro_boot.o
 INTERFACES=
